+++ fmi-base/src/main/kotlin/no/ntnu/ihb/fmi4j/xml/variables/Causality.kt
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package no.ntnu.ihb.fmi4j.modeldescription.variables
+package no.ntnu.ihb.fmi4j.xml.variables
 
 
 /**
