+++ fmi-base/src/main/kotlin/no/ntnu/ihb/fmi4j/xml/variables/ToolAnnotation.kt
@@ -1,4 +1,4 @@
-package no.ntnu.ihb.fmi4j.modeldescription.variables
+package no.ntnu.ihb.fmi4j.xml.variables
 
 typealias Annotations = List<ToolAnnotation>
