+++ fmi-base/src/test/kotlin/no/ntnu/ihb/fmi4j/modeldescription/gson/GsonTest.kt
@@ -1,16 +1,10 @@
 package no.ntnu.ihb.fmi4j.modeldescription.gson
 
-import com.google.gson.ExclusionStrategy
-import com.google.gson.FieldAttributes
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import no.ntnu.ihb.fmi.fmi2.xml.Fmi2ModelDescription
-import no.ntnu.ihb.fmi4j.modeldescription.ModelDescriptionImpl
 import no.ntnu.ihb.fmi4j.modeldescription.TestFMUs
+import no.ntnu.ihb.fmi4j.xml.ModelDescriptionImpl
 import org.junit.jupiter.api.Test
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import javax.xml.datatype.XMLGregorianCalendar
 
 class GsonTest {
