+++ fmi-base/src/test/kotlin/no/ntnu/ihb/fmi4j/modeldescription/vendors/twentysim/ControlledTemperatureTest.kt
@@ -25,9 +25,9 @@
 package no.ntnu.ihb.fmi4j.modeldescription.vendors.twentysim
 
 import no.ntnu.ihb.fmi4j.modeldescription.TestFMUs
-import no.ntnu.ihb.fmi4j.modeldescription.misc.DefaultExperiment
-import no.ntnu.ihb.fmi4j.modeldescription.variables.ModelVariables
-import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable
+import no.ntnu.ihb.fmi4j.xml.misc.DefaultExperiment
+import no.ntnu.ihb.fmi4j.xml.variables.ModelVariables
+import no.ntnu.ihb.fmi4j.xml.variables.RealVariable
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
