+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/AbstractFmuInstance.kt
@@ -27,7 +27,7 @@ package no.ntnu.ihb.fmi4j.importer
 import no.ntnu.ihb.fmi4j.common.*
 import no.ntnu.ihb.fmi4j.importer.jni.Fmi2LibraryWrapper
 import no.ntnu.ihb.fmi4j.importer.misc.FmuVariableAccessorImpl
-import no.ntnu.ihb.fmi4j.modeldescription.CommonModelDescription
+import no.ntnu.ihb.fmi4j.xml.CommonModelDescription
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
