+++ fmi-import/src/test/java/no/ntnu/ihb/fmi4j/importer/ControlledTemperatureTestJava.java
@@ -26,7 +26,7 @@ package no.ntnu.ihb.fmi4j.importer;
 
 import no.ntnu.ihb.fmi4j.common.FmiStatus;
 import no.ntnu.ihb.fmi4j.common.FmuSlave;
-import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable;
+import no.ntnu.ihb.fmi4j.xml.variables.RealVariable;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
