+++ fmi-import/src/test/java/no/ntnu/ihb/fmi4j/importer/VanDerPolTestJava.java
@@ -27,9 +27,9 @@ package no.ntnu.ihb.fmi4j.importer;
 import no.ntnu.ihb.fmi4j.common.FmiStatus;
 import no.ntnu.ihb.fmi4j.common.FmuRead;
 import no.ntnu.ihb.fmi4j.common.FmuSlave;
-import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable;
 import no.ntnu.ihb.fmi4j.solvers.Solver;
 import no.ntnu.ihb.fmi4j.solvers.apache.ApacheSolvers;
+import no.ntnu.ihb.fmi4j.xml.variables.RealVariable;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
