+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/misc/Benchmark.kt
@@ -1,8 +1,8 @@
 package no.ntnu.ihb.fmi4j.importer.misc
 
-import no.ntnu.ihb.fmi4j.importer.TestFMUs
 import no.ntnu.ihb.fmi4j.importer.Fmu
-import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable
+import no.ntnu.ihb.fmi4j.importer.TestFMUs
+import no.ntnu.ihb.fmi4j.xml.variables.RealVariable
 import org.javafmi.wrapper.Simulation
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
